@@ -1,6 +1,6 @@
 {
   "name": "@angular/cdk",
-  "version": "2.0.0-beta.8-1807c5b",
+  "version": "2.0.0-beta.8-0113c38",
   "description": "Angular Material Component Development Kit",
   "main": "./bundles/cdk.umd.js",
   "module": "./@angular/cdk.es5.js",
@@ -23,8 +23,8 @@
   },
   "homepage": "https://github.com/angular/material2#readme",
   "peerDependencies": {
-    "@angular/core": "^4.0.0",
-    "@angular/common": "^4.0.0"
+    "@angular/core": "^4.3.0",
+    "@angular/common": "^4.3.0"
   },
   "dependencies": {
     "tslib": "^1.7.1"
