@@ -749,8 +749,13 @@ class ListKeyManager {
         this._items = _items;
         this._activeItemIndex = -1;
         this._wrap = false;
-        this._nonNavigationKeyStream = new Subject();
-        this._pressedInputKeys = [];
+        this._letterKeyStream = new Subject();
+        this._pressedLetters = [];
+        /**
+         * Stream that emits any time the TAB key is pressed, so components can react
+         * when focus is shifted off of the list.
+         */
+        this.tabOut = new Subject();
     }
     /**
      * Turns on wrapping mode, which ensures that the active item will wrap to
@@ -776,12 +781,11 @@ class ListKeyManager {
         // Debounce the presses of non-navigational keys, collect the ones that correspond to letters
         // and convert those letters back into a string. Afterwards find the first item that starts
         // with that string and select it.
-        this._typeaheadSubscription = RxChain.from(this._nonNavigationKeyStream)
-            .call(filter, keyCode => keyCode >= A && keyCode <= Z)
-            .call(doOperator, keyCode => this._pressedInputKeys.push(keyCode))
+        this._typeaheadSubscription = RxChain.from(this._letterKeyStream)
+            .call(doOperator, keyCode => this._pressedLetters.push(keyCode))
             .call(debounceTime, debounceInterval)
-            .call(filter, () => this._pressedInputKeys.length > 0)
-            .call(map, () => String.fromCharCode(...this._pressedInputKeys))
+            .call(filter, () => this._pressedLetters.length > 0)
+            .call(map, () => this._pressedLetters.join(''))
             .subscribe(inputString => {
             const /** @type {?} */ items = this._items.toArray();
             for (let /** @type {?} */ i = 0; i < items.length; i++) {
@@ -790,7 +794,7 @@ class ListKeyManager {
                     break;
                 }
             }
-            this._pressedInputKeys = [];
+            this._pressedLetters = [];
         });
         return this;
     }
@@ -816,13 +820,22 @@ class ListKeyManager {
             case UP_ARROW:
                 this.setPreviousItemActive();
                 break;
-            // Note that we return here, in order to avoid preventing
-            // the default action of unsupported keys.
+            case TAB:
+                this.tabOut.next();
+                return;
             default:
-                this._nonNavigationKeyStream.next(event.keyCode);
+                if (event.keyCode >= A && event.keyCode <= Z) {
+                    // Attempt to use the `event.key` which also maps it to the user's keyboard language,
+                    // otherwise fall back to `keyCode` and `fromCharCode` which always resolve to English.
+                    this._letterKeyStream.next(event.key ?
+                        event.key.toLocaleUpperCase() :
+                        String.fromCharCode(event.keyCode));
+                }
+                // Note that we return here, in order to avoid preventing
+                // the default action of non-navigational keys.
                 return;
         }
-        this._pressedInputKeys = [];
+        this._pressedLetters = [];
         event.preventDefault();
     }
     /**
@@ -875,14 +888,6 @@ class ListKeyManager {
      */
     updateActiveItemIndex(index) {
         this._activeItemIndex = index;
-    }
-    /**
-     * Observable that emits any time the TAB key is pressed, so components can react
-     * when focus is shifted off of the list.
-     * @return {?}
-     */
-    get tabOut() {
-        return filter.call(this._nonNavigationKeyStream, keyCode => keyCode === TAB);
     }
     /**
      * This method sets the active item, given a list of items and the delta between the
